dbr/requirements-development.txt
coverage==4.3.4
django-nose==1.4.4
# mock needed for python 2.7
mock==2.0.0
model-mommy==1.3.2
nose==1.3.7
pylint==1.6.4
pep8==1.7.0
